get rid of double datatype

+++ comm/isender.h
@@ -25,11 +25,11 @@ class ISender
 {
 public:
     virtual ~ISender() {}
-    virtual void cc(int chan, int voiceId, int cc, double v1)=0;
+    virtual void cc(int chan, int voiceId, int cc, float v1)=0;
     virtual void pc(int chan, int v1)=0;
-    virtual void noteOn(int chan, int voiceId, double f, int midinote, int pitch, int v)=0;
+    virtual void noteOn(int chan, int voiceId, float f, int midinote, int pitch, int v)=0;
     virtual void noteOff(int chan, int voiceId)=0;
-    virtual void pitch(int chan, int voiceId, double f, int midinote, int pitch)=0;
+    virtual void pitch(int chan, int voiceId, float f, int midinote, int pitch)=0;
     virtual void setDestination(QHostAddress,int)=0;
 };
