+++ comm/sendermobilesynth.cpp
@@ -11,7 +11,7 @@ SenderMobileSynth::SenderMobileSynth(RC1 * rc1)
     syco->setController(sy);
 }
 
-void SenderMobileSynth::cc(int chan, int voiceId, int cc, double v1)
+void SenderMobileSynth::cc(int chan, int voiceId, int cc, float v1)
 {
 //    qDebug() << "SenderMobileSynth::cc " << voiceId << " cc " << cc << " v1 " << v1;
 
@@ -114,7 +114,7 @@ void SenderMobileSynth::pc(int chan, int v1)
     } */
 }
 
-void SenderMobileSynth::noteOn(int chan, int voiceId, double f, int midinote, int pitch, int vel)
+void SenderMobileSynth::noteOn(int chan, int voiceId, float f, int midinote, int pitch, int vel)
 {
     sy->NoteOn(voiceId, (float)f);
 }
@@ -124,7 +124,7 @@ void SenderMobileSynth::noteOff(int chan,int voiceId)
     sy->NoteOff(voiceId);
 }
 
-void SenderMobileSynth::pitch(int chan, int voiceId, double f, int midinote, int pitch) {
+void SenderMobileSynth::pitch(int chan, int voiceId, float f, int midinote, int pitch) {
     sy->NoteOn(voiceId, (float)f);
 }
 
@@ -141,7 +141,7 @@ SenderMobileSynth::SenderMobileSynth(RC1 * rc1)
     //qDebug() << "mobileSynthQt52 created";
 }
 
-void SenderMobileSynth::cc(int chan, int voiceId, int cc, double v1)
+void SenderMobileSynth::cc(int chan, int voiceId, int cc, float v1)
 {
     if(cc==2) {
 //        qDebug() << "SenderMobileSynth::cc " << voiceId << " cc " << cc << " v1 " << v1;
@@ -185,7 +185,7 @@ void SenderMobileSynth::pc(int chan, int v1)
 {
 }
 
-void SenderMobileSynth::noteOn(int chan, int voiceId, double f, int midinote, int pitch, int vel)
+void SenderMobileSynth::noteOn(int chan, int voiceId, float f, int midinote, int pitch, int vel)
 {
     sy->noteOn(voiceId, (float)f);
 }
@@ -195,7 +195,7 @@ void SenderMobileSynth::noteOff(int chan, int voiceId)
     sy->noteOff(voiceId);
 }
 
-void SenderMobileSynth::pitch(int chan, int voiceId, double f, int midinote, int pitch) {
+void SenderMobileSynth::pitch(int chan, int voiceId, float f, int midinote, int pitch) {
     sy->noteOn(voiceId, (float)f);
 }
