+++ comm/sendermulti.cpp
@@ -7,7 +7,7 @@ SenderMulti::SenderMulti(RC1 * rc1)
 
 }
 
-void SenderMulti::cc(int chan, int voiceId, int cc, double v1)
+void SenderMulti::cc(int chan, int voiceId, int cc, float v1)
 {
     s1->cc(chan,voiceId,cc,v1);
     s2->cc(chan,voiceId,cc,v1);
@@ -19,7 +19,7 @@ void SenderMulti::pc(int chan, int v1)
     s2->pc(chan,v1);
 }
 
-void SenderMulti::noteOn(int chan, int voiceId, double f, int midinote, int pitch, int v)
+void SenderMulti::noteOn(int chan, int voiceId, float f, int midinote, int pitch, int v)
 {
     s1->noteOn(chan,voiceId,f,midinote,pitch,v);
     s2->noteOn(chan,voiceId,f,midinote,pitch,v);
@@ -31,7 +31,7 @@ void SenderMulti::noteOff(int chan, int voiceId)
     s2->noteOff(chan,voiceId);
 }
 
-void SenderMulti::pitch(int chan, int voiceId, double f, int midinote, int pitch)
+void SenderMulti::pitch(int chan, int voiceId, float f, int midinote, int pitch)
 {
     s1->pitch(chan,voiceId,f,midinote,pitch);
     s2->pitch(chan,voiceId,f,midinote,pitch);
