+++ comm/senderoscpuredata.h
@@ -27,11 +27,11 @@ class SenderOscPuredata : public ISender
 {
 public:
     SenderOscPuredata(RC1 * rc1);
-    virtual void cc(int chan, int voiceId, int cc, double v1);
+    virtual void cc(int chan, int voiceId, int cc, float v1);
     virtual void pc(int chan, int v1);
-    virtual void noteOn(int chan, int voiceId, double f, int midinote, int pitch, int vel);
+    virtual void noteOn(int chan, int voiceId, float f, int midinote, int pitch, int vel);
     virtual void noteOff(int chan, int voiceId);
-    virtual void pitch(int chan, int voiceId, double f, int midinote, int pitch) {}
+    virtual void pitch(int chan, int voiceId, float f, int midinote, int pitch) {}
     virtual void setDestination(QHostAddress a,int p);
 
 private:
