+++ comm/senderoscxy.h
@@ -27,18 +27,18 @@ class SenderOscXY : public ISender
 {
 public:
     SenderOscXY(RC1 * rc1);
-    virtual void cc(int chan, int voiceId, int cc, double v1);
+    virtual void cc(int chan, int voiceId, int cc, float v1);
     virtual void pc(int chan, int v1);
-    virtual void noteOn(int chan, int voiceId, double f, int midinote, int pitch, int vel);
-    virtual void noteOff(int chan, int voiceId, double f, int midinote, int pitch, int vel);
-    virtual void pitch(int chan, int voiceId, double f, int midinote, int pitch) {}
+    virtual void noteOn(int chan, int voiceId, float f, int midinote, int pitch, int vel);
+    virtual void noteOff(int chan, int voiceId, float f, int midinote, int pitch, int vel);
+    virtual void pitch(int chan, int voiceId, float f, int midinote, int pitch) {}
     virtual void setDestination(QHostAddress a,int p);
 
 private:
     QOscClientInterface* oscout;
     RC1 * rc1;
-    double x;
-    double y;
+    float x;
+    float y;
     void sendOsc(QString path, QVariant list);
 };
